improves the code based on PR comments
