+++ src/amclient/am-policy-decision-request.ts
@@ -1,0 +1,5 @@
+export interface AmPolicyDecisionRequest {
+  resources: string[];
+  application: string;
+  subject: { ssoToken: string };
+}
