+++ src/amclient/am-policy-decision.ts
@@ -1,0 +1,6 @@
+export interface AmPolicyDecision {
+  resource: string;
+  actions: { [ action: string ]: boolean; };
+  attributes: { [ attribute: string ]: string[]; };
+  advices: any;
+}
