+++ src/amclient/am-server-info.ts
@@ -1,0 +1,4 @@
+export interface AmServerInfo {
+  cookieName: string;
+  domains: string[];
+}
