+++ src/index.ts
@@ -1,16 +1,20 @@
-export * from './am-client';
+export * from './amclient/am-client';
+export * from './amclient/am-policy-decision';
+export * from './amclient/am-policy-decision-request';
+export * from './amclient/am-server-info';
 export * from './cache/cache';
 export * from './cache/in-memory-cache';
 export * from './error/invalid-session-error';
 export * from './error/shield-evaluation-error';
-export * from './policy-agent';
-export * from './policy-agent-options';
+export * from './policyagent/policy-agent';
+export * from './policyagent/policy-agent-options';
 export * from './shield/basic-auth-shield';
 export * from './shield/cookie-shield';
 export * from './shield/oauth2-shield';
 export * from './shield/policy-shield';
 export * from './shield/session-data';
 export * from './shield/shield';
+export * from './utils/deferred';
 export * from './utils/http-utils';
 export * from './utils/logger';
 export * from './utils/xml-utils';
