+++ src/shield/cookie-shield.ts
@@ -1,7 +1,7 @@
 import { IncomingMessage, ServerResponse } from 'http';
 
 import { ShieldEvaluationError } from '../error/shield-evaluation-error';
-import { PolicyAgent } from '../policy-agent';
+import { PolicyAgent } from '../policyagent/policy-agent';
 import { Deferred } from '../utils/deferred';
 import { redirect } from '../utils/http-utils';
 import { SessionData } from './session-data';
